Moved processCommand to raceSequence as it has logic based on raceState
